[breaking] Update pdfjs-dist to 4.0.379
