+++ packages/react-pdf/package.json
@@ -60,7 +60,7 @@
     "make-cancellable-promise": "^1.3.1",
     "make-event-props": "^1.6.0",
     "merge-refs": "^1.3.0",
-    "pdfjs-dist": "3.11.174",
+    "pdfjs-dist": "4.0.379",
     "tiny-invariant": "^1.0.0",
     "warning": "^4.0.0"
   },
