+++ packages/react-pdf/src/Page/AnnotationLayer.tsx
@@ -5,7 +5,7 @@ import makeCancellable from 'make-cancellable-promise';
 import clsx from 'clsx';
 import invariant from 'tiny-invariant';
 import warning from 'warning';
-import pdfjs from '../pdfjs.js';
+import * as pdfjs from 'pdfjs-dist';
 
 import useDocumentContext from '../shared/hooks/useDocumentContext.js';
 import usePageContext from '../shared/hooks/usePageContext.js';
