+++ packages/react-pdf/src/Page/Canvas.spec.tsx
@@ -1,10 +1,9 @@
-import { Blob } from 'node:buffer';
-import { beforeAll, describe, expect, it } from 'vitest';
+import { beforeAll, describe, expect, it, vi } from 'vitest';
 import { render } from '@testing-library/react';
 
 import { pdfjs } from '../index.test.js';
 
-import PageSVG from './PageSVG.js';
+import Canvas from './Canvas.js';
 
 import failingPage from '../../../../__mocks__/_failing_page.js';
 
@@ -33,28 +32,35 @@ function renderWithContext(children: React.ReactNode, context: Partial<PageConte
   };
 }
 
-describe('PageSVG', () => {
+describe('Canvas', () => {
   // Loaded page
   let page: PDFPageProxy;
+  let pageWithRendererMocked: PDFPageProxy;
 
   beforeAll(async () => {
     const pdf = await pdfjs.getDocument({ data: pdfFile.arrayBuffer }).promise;
 
     page = await pdf.getPage(1);
+
+    pageWithRendererMocked = Object.assign(page, {
+      render: () => ({
+        promise: new Promise<void>((resolve) => resolve()),
+        cancel: () => {
+          // Intentionally empty
+        },
+      }),
+    });
   });
 
   describe('loading', () => {
     it('renders a page and calls onRenderSuccess callback properly', async () => {
-      const originalBlob = globalThis.Blob;
-      globalThis.Blob = Blob as unknown as typeof globalThis.Blob;
-
       const { func: onRenderSuccess, promise: onRenderSuccessPromise } = makeAsyncCallback();
 
       muteConsole();
 
-      renderWithContext(<PageSVG />, {
+      renderWithContext(<Canvas />, {
         onRenderSuccess,
-        page,
+        page: pageWithRendererMocked,
         scale: 1,
       });
 
@@ -63,8 +69,6 @@ describe('PageSVG', () => {
       await expect(onRenderSuccessPromise).resolves.toMatchObject([{}]);
 
       restoreConsole();
-
-      globalThis.Blob = originalBlob;
     });
 
     it('calls onRenderError when failed to render canvas', async () => {
@@ -72,7 +76,7 @@ describe('PageSVG', () => {
 
       muteConsole();
 
-      renderWithContext(<PageSVG />, {
+      renderWithContext(<Canvas />, {
         onRenderError,
         page: failingPage,
         scale: 1,
@@ -83,6 +87,55 @@ describe('PageSVG', () => {
       await expect(onRenderErrorPromise).resolves.toMatchObject([expect.any(Error)]);
 
       restoreConsole();
+    });
+  });
+
+  describe('rendering', () => {
+    it('passes canvas element to canvasRef properly', () => {
+      const canvasRef = vi.fn();
+
+      renderWithContext(<Canvas canvasRef={canvasRef} />, {
+        page: pageWithRendererMocked,
+        scale: 1,
+      });
+
+      expect(canvasRef).toHaveBeenCalled();
+      expect(canvasRef).toHaveBeenCalledWith(expect.any(HTMLElement));
+    });
+
+    it('does not request structure tree to be rendered when renderTextLayer = false', async () => {
+      const { func: onRenderSuccess, promise: onRenderSuccessPromise } = makeAsyncCallback();
+
+      const { container } = renderWithContext(<Canvas />, {
+        onRenderSuccess,
+        page: pageWithRendererMocked,
+        renderTextLayer: false,
+      });
+
+      await onRenderSuccessPromise;
+
+      const structTree = container.querySelector('.react-pdf__Page__structTree');
+
+      expect(structTree).not.toBeInTheDocument();
+    });
+
+    it('renders StructTree when given renderTextLayer = true', async () => {
+      const { func: onGetStructTreeSuccess, promise: onGetStructTreeSuccessPromise } =
+        makeAsyncCallback();
+
+      const { container } = renderWithContext(<Canvas />, {
+        onGetStructTreeSuccess,
+        page: pageWithRendererMocked,
+        renderTextLayer: true,
+      });
+
+      expect.assertions(1);
+
+      await onGetStructTreeSuccessPromise;
+
+      const canvas = container.querySelector('canvas') as HTMLCanvasElement;
+
+      expect(canvas.children.length).toBeGreaterThan(0);
     });
   });
 });
