--- packages/react-pdf/src/pdfjs.ts
@@ -1,7 +1,0 @@
-import * as pdfjsModule from 'pdfjs-dist';
-
-const pdfjs = (
-  'default' in pdfjsModule ? pdfjsModule['default'] : pdfjsModule
-) as typeof pdfjsModule;
-
-export default pdfjs;
