+++ packages/react-pdf/tsconfig.json
@@ -10,7 +10,7 @@
     "outDir": "dist",
     "skipLibCheck": true,
     "strict": true,
-    "target": "es2015",
+    "target": "es2017",
     "verbatimModuleSyntax": true
   },
   "exclude": ["dist"]
