+++ sample/create-react-app-5/.gitignore
@@ -3,4 +3,4 @@ build
 node_modules
 public/cmaps
 public/standard_fonts
-public/pdf.worker.js
+public/pdf.worker.mjs
