+++ sample/create-react-app-5/scripts/copy-worker.ts
@@ -5,10 +5,10 @@ import { createRequire } from 'node:module';
 const require = createRequire(import.meta.url);
 
 const pdfjsDistPath = path.dirname(require.resolve('pdfjs-dist/package.json'));
-const pdfWorkerPath = path.join(pdfjsDistPath, 'build', 'pdf.worker.js');
+const pdfWorkerPath = path.join(pdfjsDistPath, 'build', 'pdf.worker.mjs');
 
 const targetDir = 'public';
-const targetPath = path.join(targetDir, 'pdf.worker.js');
+const targetPath = path.join(targetDir, 'pdf.worker.mjs');
 
 // Ensure target directory exists
 fs.mkdirSync(targetDir, { recursive: true });
