+++ sample/vite/vite.config.ts
@@ -13,6 +13,14 @@ const standardFontsDir = normalizePath(
 );
 
 export default defineConfig({
+  build: {
+    target: 'es2022',
+  },
+  optimizeDeps: {
+    esbuildOptions: {
+      target: 'es2022',
+    },
+  },
   plugins: [
     viteStaticCopy({
       targets: [
