+++ test/vite.config.ts
@@ -15,6 +15,14 @@ const standardFontsDir = normalizePath(
 
 export default defineConfig({
   base: './',
+  build: {
+    target: 'es2022',
+  },
+  optimizeDeps: {
+    esbuildOptions: {
+      target: 'es2022',
+    },
+  },
   plugins: [
     react(),
     viteStaticCopy({
